+++ src/memory.c
@@ -1,10 +1,10 @@
 #include "memory.h"
 
-void* xmalloc(uint64_t size) {
+void* xmalloc(size_t size) {
     return malloc(size);
 }
 
-void* xrealloc(void* ptr, uint64_t new_size) {
+void* xrealloc(void* ptr, size_t new_size) {
     return realloc(ptr, new_size);
 }
