+++ src/memory.h
@@ -2,10 +2,10 @@
 #define _CIR_MEMORY_H_
 
 #include <stdlib.h>
-#include <stdint.h>
+#include <stddef.h>
 
-void* xmalloc(uint64_t size);
-void* xrealloc(void* ptr, uint64_t new_size);
+void* xmalloc(size_t size);
+void* xrealloc(void* ptr, size_t new_size);
 void xfree(void* ptr);
 
 #endif
